feat(demo): add end-to-end execution pipeline
- Created `run_pipeline.py` to demonstrate the full system workflow.
- Integrates `RiskPredictionModel` (LSTM), `InterventionSimulator` (Seq2Seq), and `ActorCritic` (RL Agent).
- Implements a complete "Patient Journey": Diagnosis -> AI Prescription -> Simulation -> Re-evaluation.
- Provides clear CLI output showing Risk Scores, Selected Treatments, and Predicted Outcomes.

+++ ml-services/intervention-simulation/src/rl_trainer.py
@@ -148,7 +148,7 @@ def train():
             state_n, reward, done, _ = trainer.env.step(action[0], action[1])
 
             # Save data for training
-            memory.states.append(torch.FloatTensor(state))
+            memory.states.append(state)
             memory.actions_cat.append(torch.tensor(action[0]))
             memory.actions_cont.append(torch.tensor(action[1]))
             # Recalculate logprob inside update loop usually, but here we store for simplicity
